refactor: :recycle: Refactor & Self CR 2.
