+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/fields/AllMonthDays/AllMonthDays.types.ts
@@ -1,0 +1,10 @@
+import { FieldProps, LeadingZero, SetValueNumbersOrUndefined } from '../../types';
+
+export interface AllMonthDaysProps extends Omit<FieldProps, 'setValue'> {
+    humanizeLabels: boolean;
+    monthDays: number[];
+    setMonthDays: SetValueNumbersOrUndefined;
+    nthMonthDays: number[];
+    setNthMonthDays: SetValueNumbersOrUndefined;
+    leadingZero: LeadingZero;
+}
