+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/fields/AllMonthDays/MonthDays/MonthDays.types.ts
@@ -1,0 +1,7 @@
+import { FieldProps, LeadingZero } from '#src-app/components/cron/types';
+
+export interface MonthDaysProps extends FieldProps {
+    humanizeLabels: boolean;
+    weekDays?: number[];
+    leadingZero: LeadingZero;
+}
