+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/fields/AllMonthDays/MonthDays/index.ts
@@ -1,0 +1,3 @@
+import MonthDays from './MonthDays';
+
+export default MonthDays;
