+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/fields/AllMonthDays/NthMonthDays/NthMonthDays.types.ts
@@ -1,0 +1,6 @@
+import { FieldProps, SetValueNumbersOrUndefined } from '#src-app/components/cron/types';
+
+export interface NthMonthDaysProps extends Omit<FieldProps, 'setValue'> {
+    humanizeLabels: boolean;
+    setValue: SetValueNumbersOrUndefined;
+}
