+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/fields/AllMonthDays/NthMonthDays/index.ts
@@ -1,0 +1,3 @@
+import NthMonthDays from './NthMonthDays';
+
+export default NthMonthDays;
