+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/fields/AllWeekDays/NthWeekDays/NthWeekDays.types.ts
@@ -1,0 +1,8 @@
+import { FieldProps, SetValueNumbersOrUndefined } from '#src-app/components/cron/types';
+
+export interface NthWeekDaysProps extends Omit<FieldProps, 'setValue'> {
+    humanizeLabels: boolean;
+    setValue: SetValueNumbersOrUndefined;
+    isAnyWeekDaySelected: boolean;
+    isMonthPeriodDisplayed: boolean;
+}
