+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/fields/AllWeekDays/NthWeekDays/index.ts
@@ -1,0 +1,3 @@
+import NthWeekDays from './NthWeekDays';
+
+export default NthWeekDays;
