+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/fields/AllWeekDays/WeekDays/index.ts
@@ -1,0 +1,3 @@
+import WeekDays from './WeekDays';
+
+export default WeekDays;
