+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/fields/Hours/Hours.types.ts
@@ -1,0 +1,6 @@
+import { ClockFormat, FieldProps, LeadingZero } from '../../types';
+
+export interface HoursProps extends FieldProps {
+    leadingZero: LeadingZero;
+    clockFormat?: ClockFormat;
+}
