+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/fields/Hours/index.ts
@@ -1,0 +1,3 @@
+import Hours from './Hours';
+
+export default Hours;
