+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/fields/Minutes/Minutes.types.ts
@@ -1,0 +1,7 @@
+import { ClockFormat, FieldProps, LeadingZero } from '../../types';
+
+
+export interface MinutesProps extends FieldProps {
+    leadingZero: LeadingZero;
+    clockFormat?: ClockFormat;
+}
