+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/fields/Minutes/index.ts
@@ -1,0 +1,3 @@
+import Minutes from './Minutes';
+
+export default Minutes;
