+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/fields/Months/Months.types.ts
@@ -1,0 +1,5 @@
+import { FieldProps } from '../../types';
+
+export interface MonthsProps extends FieldProps {
+    humanizeLabels: boolean;
+}
