+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/cron/fields/Months/index.ts
@@ -1,0 +1,3 @@
+import Months from './Months';
+
+export default Months;
