+++ runbotics/runbotics-orchestrator-ui/src/src-app/translations/en/components.json
@@ -1,6 +1,6 @@
 {
     "Component.SettingsNotification.Title": "Settings Updated",
-    "Component.SettingsNofigication.Message": "We automatically updated your settings. You change the settings any time from your dashboard settings.",
+    "Component.SettingsNotification.Message": "We automatically updated your settings. You change the settings any time from your dashboard settings.",
 
     "Component.GenericMoreOptions.MoreOptions": "More options",
