+++ runbotics/runbotics-orchestrator-ui/src/src-app/translations/pl/components.json
@@ -1,6 +1,6 @@
 {
     "Component.SettingsNotification.Title": "Ustawienia zaktualizowane",
-    "Component.SettingsNofigication.Message": "Automatycznie zaktualizowaliśmy Twoje ustawienia. Możesz zmienić ustawienia w dowolnym momencie w ustawieniach pulpitu.",
+    "Component.SettingsNotification.Message": "Automatycznie zaktualizowaliśmy Twoje ustawienia. Możesz zmienić ustawienia w dowolnym momencie w ustawieniach pulpitu.",
 
     "Component.GenericMoreOptions.MoreOptions": "Więcej opcji",
